Added permission checks to API

+++ src/routes/api/cards/+server.ts
@@ -1,8 +1,8 @@
 // src/routes/api/cards/+server.ts
 import type { RequestHandler } from './$types';
-import { StoredCard, type CardID } from '$lib/domain/cards/cardStore.svelte';
 import type { card as PrismaCard } from '@prisma/client';
 import { db } from '$lib/server/db';
+import { MAX_CARDS } from '$lib/server/rules/maxCards';
 
 export type CardPayload = {
 	action: 'create' | 'update' | 'delete';
@@ -23,10 +23,18 @@ export type CardPayload = {
 export const GET: RequestHandler = async ({ locals }) => {
 	const user = locals.user;
 	if (!user) return new Response('Unauthorized, client is not a logged in user', { status: 401 });
-	// TODO: add to schema and uncomment
-	// if (user.canView !== true) return new Response('Unauthorized', { status: 401 });
 
-	const cards = await db.card.findMany(); // TODO: add filter for user
+	// Get all cards the user can see
+	const cards = await db.card.findMany({
+		where: {
+			OR: [
+				{ ownerId: user.id },
+				{ editors: { some: { id: user.id } } },
+				{ viewers: { some: { id: user.id } } },
+				{ public: true }
+			]
+		}
+	}); // TODO: add filter for user
 	return new Response(JSON.stringify({ success: true, cards }), { status: 200 });
 };
 
@@ -40,8 +48,16 @@ export const POST: RequestHandler = async ({ locals, request }) => {
 	// ✅ ensure user is logged in
 	const user = locals.user;
 	if (!user) return new Response('Unauthorized, client is not a logged in user', { status: 401 });
-	// TODO: add to schema and uncomment
-	// if (user.canCreate !== true) return new Response('Unauthorized', { status: 401 });
+
+	// Give user a max amount of cards to create
+	const dbUser = await db.user.findUniqueOrThrow({
+		where: { id: user.id },
+		include: { ownedCards: true }
+	});
+	if (dbUser.ownedCards.length >= MAX_CARDS)
+		return new Response('Unauthorized, user has reached the maximum amount of cards', {
+			status: 401
+		});
 
 	// Get and prepare request
 	let prismaCard: PrismaCard = await request.json();
@@ -81,8 +97,15 @@ export const PUT: RequestHandler = async ({ locals, request }) => {
 	prismaCard.updatedAt = new Date(); // Override updatedAt (fallback)
 
 	// Check if the card already exists
-	const exists = await db.card.findUnique({ where: { id: prismaCard.id } });
-	if (!exists) return new Response('Card does not exist', { status: 400 });
+	const dbCard = await db.card.findUnique({
+		where: { id: prismaCard.id },
+		include: { owner: true, editors: true }
+	});
+	if (!dbCard) return new Response('Card does not exist', { status: 400 });
+
+	// Check if user is authorized to update card (editor or owner)
+	if (!(user.id === dbCard.ownerId || dbCard.editors.some((editor) => editor.id === user.id)))
+		return new Response('Unauthorized, card belongs to another user', { status: 401 });
 
 	// Update in DB
 	await db.card.update({
@@ -104,13 +127,8 @@ export const DELETE: RequestHandler = async ({ locals, request }) => {
 	// ✅ ensure user is logged in
 	const user = locals.user;
 	if (!user) return new Response('Unauthorized, client is not a logged in user', { status: 401 });
-	// TODO: add to schema and uncomment
-	// if (user.canDelete !== true) return new Response('Unauthorized', { status: 401 });
 
 	let prismaCards: PrismaCard[] = await request.json();
-	// Check if the user is allowed to delete this card
-	// TODO: add to schema and uncomment
-	// if (user.isAdmin !== true && prismaCard.ownerId !== user.id) return new Response('Unauthorized', { status: 401 });
 
 	const uniqueIds = prismaCards.map((card) => card.id);
 	const cardsToDelete = await db.card.findMany({
@@ -123,6 +141,11 @@ export const DELETE: RequestHandler = async ({ locals, request }) => {
 			'Unauthorized, some cards does not belong to user or do not exist in database',
 			{ status: 401 }
 		);
+
+	// TODO: add admin override
+	// Check if user is owner (allowed to delete)
+	if (cardsToDelete.some((card) => card.ownerId !== user.id))
+		return new Response('Unauthorized, some cards do not belong to user', { status: 401 });
 
 	// Delete in DB
 	await db.card.deleteMany({ where: { id: { in: uniqueIds }, ownerId: user.id } });
